Refactor base processor class to be more generic
