+++ src/mmore/process/processors/docx_processor.py
@@ -5,7 +5,7 @@ from typing import List
 from PIL import Image
 from src.mmore.process.utils import clean_text
 from src.mmore.type import FileDescriptor, MultimodalSample
-from .processor import Processor, ProcessorConfig
+from .base import Processor, ProcessorConfig
 from docx.opc.constants import RELATIONSHIP_TYPE as RT 
 
 logger = logging.getLogger(__name__)
