+++ src/mmore/process/processors/eml_processor.py
@@ -6,7 +6,7 @@ from PIL import Image
 from typing import Dict, Any, List
 from src.mmore.process.utils import clean_text
 from src.mmore.type import FileDescriptor, MultimodalSample
-from .processor import Processor, ProcessorConfig
+from .base import Processor, ProcessorConfig
 
 logger = logging.getLogger(__name__)
