+++ src/mmore/process/processors/md_processor.py
@@ -2,7 +2,7 @@ import logging
 import markdown
 import markdownify
 from src.mmore.type import FileDescriptor, MultimodalSample
-from .processor import Processor, ProcessorConfig
+from .base import Processor, ProcessorConfig
 import tempfile
 from PIL import Image
 import os
